Force Login, Side Menu Bar, Bootstrap Theme Integration
Used a free bootstrap theme as springboard.
Please use the theme as a CSS API (provided in links below).

We will further customize the look at view as we go in the css files.
However let's stick with the set of CSS classes and ids so that we can
modify it easily in the future.

http://startbootstrap.com/sb-admin-v2